+++ backend/jest.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  testEnvironment: 'node',
+  coveragePathIgnorePatterns: ['/node_modules/'],
+  testMatch: ['**/tests/**/*.test.js'],
+  collectCoverageFrom: [
+    '**/*.js',
+    '!**/tests/**',
+    '!**/node_modules/**',
+    '!**/coverage/**'
+  ]
+};